requirements.txt
opencv-python==4.7.0.72
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.0
matplotlib==3.7.1
albumentations==1.3.1
numpy==1.24.3
PyYAML==6.0
tensorboardX==2.6
timm==1.0.15
